Add buffering to the gauges
As in when a gauge's value is modified, the change will only
take effect on the next frame.
This means that if multiple changes are applied on the same frame
(eg. new turn), all changes will be applied together on the next frame

This is particularly usefull to avoid needlessly triggering too many
signals, and in cases such as:
The gauge is at 5, the lower limit is 0. we apply -15 and +20 to the
gauge.
Before you would end up with 20 because the -15 wouldn't be applied in
full because of the lower limit