Add easing for repeat event probability
This means that events that were recently triggered have a lower weight.

The easing currently depends on the number of other events that have
been triggered rather than time/days

This easing is parametrized in an Event's exported variables:
weight_ease_duration for the number of other triggers until the weight
comes back to normal
weight_ease_exponent for the curve of the ease function. Use 1 for a
linear function